TFRecordWriter inheritance and batch splitting
Our TFRecordWriters (single and paired images) no longer encapsulate a TFRecordWriter but are its subclasses.
In this way we can use the writers in python's 'with' blocks and let them deal with the close operation.
This is particularly useful if we want to split the images across multiple tfrecords.
(the way we stop reading images from the queue is by catching the OutOfRangeError and with the writer in
a 'with' block it closes itself automatically.